MM-57292 - Adding root post divider component
